Remove invalid test cases.
There's a misunderstanding of HTML5 canvas API.

For given width, height of the CanvasRenderingContext2D.createImageData,
those can be any values until it resides within the range of unsigned
long.

In ohter words, the width and height bigger than the size of canvas
don't mean a invalid image data.

For more information about the createImageData(), check:
http://www.w3.org/html/wg/drafts/2dcontext/html5_canvas_CR/#dom-context-2d-createimagedata

This patch removes that wrong test case.